backend/requirements.txt
# ============================================================================
# FastAPI Backend Dependencies
# AgriVisionTalk - Multimodal AI for Smart Farming
# ============================================================================

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
# psycopg2-binary==2.9.9  # Not needed for SQLite testing
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2

# AI/ML
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
scikit-learn==1.3.0

# HuggingFace & RAG
transformers==4.37.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
langchain==0.1.6
langchain-community==0.0.19

# Weather API
requests==2.31.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
pydantic-settings==2.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Logging & Monitoring
python-json-logger==2.0.7

# CORS & Middleware
python-jose[cryptography]==3.3.0
